Use Binomial Distribution
Updated point cloud generation to use a uniformly distributed set of points over each 2010 census block with weights from a binomial distribution. These weights are multiplied by the population of a 2010 census block in order to distribute the population among points in the block while weighting for distance from the block's centroid.