Compare Population Weighted and Smooth Surface Interpolation
Population weighted interpolation outperforms smooth surface interpolation by approximately 1/100 of a percent. Since population weighted interpolation is much faster to perform and is a deterministic algorithm, it is the preferred method for most cases. However, the opportunity to use tertiary data to reshape the smooth surface data is promising for increased accuracy. It is also noteworthy that the smooth surface process may be faster when you want to map a dataset to a large variety of regions.